Apply same namespacing pattern to ccr as to cclf
